Move login and signup middleware to authController
